+++ trackinsert/src/main/java/dwsc/trackinsert/repository/TrackRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.CrudRepository;
 
 import dwsc.trackinsert.model.Track;
 
-public interface TrackRepository extends CrudRepository<Track, Long> {
+public interface TrackRepository extends CrudRepository<Track, Integer> {
 
 	<S extends Track> S save(S track);
-	Optional<Track> findById(Long id);
+	Optional<Track> findById(int id);
 
 }
